refactor: editor 내부 type 분리
